requirements.txt
numpy<2.0.0
matplotlib
lightning
torchmetrics
torchvision
torch-geometric
kornia
scipy
scikit-image
rich
dill